Add kustomize to configure paths separately for development

+++ docs/starcry-dev-wrapper.sh
@@ -5,8 +5,12 @@ echo wrapper start
 set -ex
 
 while true; do
-    cp -prv /workdir/build/starcry /starcry
-    /starcry "$@"
+    if [[ -f /workdir/build/starcry ]]; then
+        cp -prv /workdir/build/starcry /starcry
+        /starcry "$@"
+    else
+        echo no binary found yet...
+    fi
     ls -al /workdir
     sleep 5
 done
